backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.0.0
pydantic>=2.0.0
openai>=1.0.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.26.0
pytest>=7.0.0
